Move DialogState to menu state
Move DialogState from online to menu state
