+++ src/View/DialogMenu/index.tsx
@@ -1,15 +1,16 @@
 import React from "react"
 import { Dialog } from "components"
-import { online } from "state/online"
 import { useParams } from "react-router-dom"
-import { DialogState } from "state/online/state/index.types"
 import { useGState } from "state"
+import { menu } from "state/menu"
+import { online } from "state/online"
+import { DialogState } from "state/menu/state/index.types"
 import OnlineSession from "./OnlineSession"
 import InitialSelection from "./InitialSelection"
 import OnlineSessionJoinOnly from "./OnlineSessionJoinOnly"
 
 const handleClose = () => {
-    online.setSessionDialog(DialogState.Closed)
+    menu.setSessionDialog(DialogState.Closed)
 }
 
 const contents = {
@@ -21,10 +22,10 @@ const contents = {
 }
 
 const DialogMenu: React.FC = () => {
-    const { dialogState, session } = useGState("SessionDialog").online
+    const { dialogState } = useGState("DialogState").menu
     const { sessionId } = useParams()
     const contentType =
-        sessionId && !session?.isConnected()
+        sessionId && !online.state.session?.isConnected()
             ? DialogState.JoinOnly
             : dialogState
