+++ src/state/menu/state/index.types.ts
@@ -6,7 +6,16 @@ export enum MainMenuState {
     Session,
 }
 
+export enum DialogState {
+    Closed,
+    InitialSelectionFirstLoad,
+    InitialSelection,
+    CreateOnlineSession,
+    JoinOnly,
+}
+
 export interface MenuState {
+    dialogState: DialogState
     mainMenuState: MainMenuState
     shortcutsOpen: boolean
     subscribeOpen: boolean
