+++ src/state/online/state/index.ts
@@ -8,14 +8,13 @@ import {
     animals,
     uniqueNamesGenerator,
 } from "unique-names-generator"
-import { DialogState, OnlineState } from "./index.types"
+import { OnlineState } from "./index.types"
 import { GlobalState, SerializedState } from "../../types"
 import { DrawingState } from "../../drawing/state/index.types"
 import { deserializeOnlineToken, serializeOnlineState } from "../serializers"
 
 export class Online implements GlobalState<OnlineState> {
     state: OnlineState = {
-        dialogState: DialogState.InitialSelectionFirstLoad,
         userSelection: {
             alias: uniqueNamesGenerator({
                 dictionaries: [adjectives, animals],
@@ -37,12 +36,6 @@ export class Online implements GlobalState<OnlineState> {
 
     setState(newState: OnlineState): void {
         this.state = newState
-    }
-
-    setSessionDialog(dialogState: DialogState): void {
-        this.state.dialogState = dialogState
-        subscriptionState.render("SessionDialog")
-        this.saveToLocalStorage()
     }
 
     newSession(session: Session): void {
