+++ src/state/online/state/index.types.ts
@@ -1,19 +1,10 @@
 import { Session, User } from "api/types"
 
 export interface OnlineState {
-    dialogState: DialogState
     userSelection: User
     session?: Session
     token?: string
     isConnected: () => boolean
     isAuthorized: () => boolean
     isSignedIn: () => boolean
-}
-
-export enum DialogState {
-    Closed,
-    InitialSelectionFirstLoad,
-    InitialSelection,
-    CreateOnlineSession,
-    JoinOnly,
 }
