+++ src/state/subscription/index.ts
@@ -25,13 +25,13 @@ export class SubscriptionState {
         PageContent: [],
         LayerConfig: [],
         Loading: [],
+        DialogState: [],
         MainMenu: [],
         MenuPageButton: [],
         Notification: [],
         Theme: [],
         RenderNG: [],
         Session: [],
-        SessionDialog: [],
         Settings: [],
         SettingsMenu: [],
         ShortcutsOpen: [],
